+++ src/i8c/constants.py
@@ -225,3 +225,4 @@ I8_OP_load_external = 0x101
 I8_OP_deref_int = 0x102
 I8_OP_cast_int2ptr = 0x103
 I8_OP_cast_ptr2int = 0x104
+I8_OP_warn = 0x105